+++ metasvc-model/src/main/java/de/digitalcollections/model/security/User.java
@@ -6,7 +6,6 @@ import java.util.List;
 import javax.validation.constraints.Email;
 import javax.validation.constraints.NotBlank;
 import lombok.experimental.SuperBuilder;
-import org.springframework.util.ObjectUtils;
 
 /** An user of the system. */
 @SuperBuilder(buildMethodName = "prebuild")
@@ -82,7 +81,7 @@ public class User extends UniqueObject {
    * @param passwordHash new hashed password
    */
   public void setPasswordHash(String passwordHash) {
-    if (!ObjectUtils.isEmpty(passwordHash)) {
+    if (passwordHash != null && !passwordHash.isEmpty()) {
       this.passwordHash = passwordHash;
     }
   }